+++ src/Model/Category/services.php
@@ -8,9 +8,7 @@ add_action('fire/services/core', function ($fire) {
     $fire->singleton('category.repository', function ($fire) {
         $repo = new CategoryRepository(CategoryTaxonomy::NAME);
 
-        $repo->registerEntityMapper(function () {
-            return new TermEntityMapper;
-        });
+        $repo->registerEntityMapper($fire['term.entitymapper']);
 
         $repo->registerEntityMapper(function () use ($fire) {
             return new CategoryEntityMapper(