+++ src/Model/Upload/services.php
@@ -8,13 +8,7 @@ add_action('fire/services/core', function ($fire) {
     $fire->singleton('upload.repository', function ($fire) {
         $repo = new UploadRepository(UploadPostType::TYPE);
 
-        $repo->registerEntityMapper(function () use ($fire) {
-            return new AbstractPostEntityMapper(
-                $fire['user.repository'],
-                $fire['upload.repository'],
-                $fire['comment.repository']
-            );
-        });
+        $repo->registerEntityMapper($fire['abstractpost.entitymapper']);
 
         $repo->registerEntityMapper(function () use ($fire) {
             return new UploadEntityMapper;