+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Fixed bug when replacing absolute URLs inside of serialized data
 - Layout functionality is now optional
 - Upload `srcset` method now correctly passes size value to WP function
+- `Asset::url` fix for Windows path separator issue
 
 ## [2.1.1] - 2015-10-27
 ### Fixed