+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Layout functionality is now optional
 - Upload `srcset` method now correctly passes size value to WP function
 - `Asset::url` fix for Windows path separator issue
+- Typo in `Collection::jsonSerialize` method name
 
 ## [2.1.1] - 2015-10-27
 ### Fixed